% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogy_Functions.R
\name{generate_sequential_string}
\alias{generate_sequential_string}
\title{Generate Sequential String of Letters}
\usage{
generate_sequential_string(length)
}
\arguments{
\item{length}{The length of the letter string to generate.}
}
\value{
A string of sequential letters of the specified length.
}
\description{
Generates a sequential string of a specified length starting from a random letter.
}
\examples{
generate_sequential_string(5)
}
