% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogy_Functions.R
\name{letter_addition}
\alias{letter_addition}
\title{Letter Addition}
\usage{
letter_addition(string)
}
\arguments{
\item{string}{The string to which a letter will be added.}
}
\value{
The string with an additional letter appended.
}
\description{
Adds the next sequential letter to the end of the string. Wraps around to 'A' after 'Z'.
}
\examples{
letter_addition("ABC")  # Returns "ABCD"
}
