% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogy_Functions.R
\name{letter_deletion}
\alias{letter_deletion}
\title{Letter Deletion}
\usage{
letter_deletion(string)
}
\arguments{
\item{string}{The string from which to delete the last letter.}
}
\value{
The string minus the last letter.
}
\description{
Removes the last letter from the given string.
}
\examples{
letter_deletion("hi")  # Returns "h"
}
