% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analogy_Functions.R
\name{reverse_string}
\alias{reverse_string}
\title{Reverse String}
\usage{
reverse_string(string)
}
\arguments{
\item{string}{The string to reverse.}
}
\value{
The reversed string.
}
\description{
Returns a new string that is the 'mirror' of the input string.
}
\examples{
reverse_string("hello")  # Returns "olleh"
}
